extract QueryResult
