+++ clickhouse-native-jdbc/src/main/java/com/github/housepower/jdbc/statement/ClickHousePreparedInsertStatement.java
@@ -51,7 +51,7 @@ public class ClickHousePreparedInsertStatement extends AbstractPreparedStatement
         this.block = getSampleBlock(insertQuery);
         this.block.initWriteBuffer();
         this.blockInit = true;
-        new ValuesWithParametersInputFormat(fullQuery, posOfData).fillBlock(block);
+        new ValuesWithParametersInputFormat(posOfData, fullQuery).fillBlock(block);
     }
 
     @Override
