+++ clickhouse-native-jdbc/src/main/java/com/github/housepower/jdbc/stream/InputFormat.java
@@ -19,5 +19,6 @@ import com.github.housepower.jdbc.data.Block;
 import java.sql.SQLException;
 
 public interface InputFormat {
+
     void fillBlock(Block block) throws SQLException;
 }
